mcp/toolsession/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "toolsession-mcp"
version = "1.0.0"
description = "Interactive tool session management with DB persistence"
authors = [{name = "ToolSession", email = "admin@toolsession.local"}]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "mcp",
    "pydantic",
    "fastapi",
    "fastapi-mcp",
    "uvicorn",
    "motor",
    "pymongo",
    "aiohttp"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0"
]

test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "aiohttp>=3.8.0"
]

[project.scripts]
toolsession-server = "server:main"
test-python-session = "test_python_session:main"
launch-python-server = "launch_python_server:main"
test-mcp-client = "test_mcp_client:main"

[tool.setuptools]
packages = []
py-modules = [
    "toolsession_server",
    "tools",
    "mongo",
    "test_python_session",
    "launch_python_server",
    "test_mcp_client"
]

[tool.black]
line-length = 88
target-version = ["py39", "py310", "py311", "py312"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.pytest.ini_options]
minversion = "7.0"
addopts = ["-ra", "-q", "--strict-markers"]
testpaths = ["."]
python_files = ["test_*.py"]
asyncio_mode = "auto"